+++ bnd-workspace/de.fraunhofer.abm.app/src/de/fraunhofer/abm/app/controllers/CollectionController.java
@@ -97,6 +97,22 @@ public class CollectionController extends AbstractController implements REST {
         }
         return result;
     }
+    
+    interface AccountRequest extends RESTRequest {
+		Map<String, String> _body();
+	}
+    
+    public void putCollectionstatus(AccountRequest ar) throws Exception {
+  	   authorizer.requireRole("UserAdmin");
+  	   Map<String, String> params = ar._body();
+  		String collectionid = params.get("collectionid");
+          try {
+           collectionDao.activeCollection(collectionid);
+          } catch (Exception e) {
+          	logger.info("Exception");
+          }
+  }
+
 
     interface CollectionRequest extends RESTRequest {
         CollectionDTO _body();
@@ -207,6 +223,7 @@ public class CollectionController extends AbstractController implements REST {
         
         return null;
     }
+    
 
     @Override
     Logger getLogger() {
